fix(orcad): preserve degradation types after rebase
